modification strlen, split, Makefile

+++ srcs/part1/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: adenord <alexandre.denord@gmail.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/13 15:17:27 by adenord           #+#    #+#             */
-/*   Updated: 2023/07/26 11:46:17 by adenord          ###   ########.fr       */
+/*   Updated: 2023/10/17 09:18:30 by adenord          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
